Use explicit cast to convert integers to enum
Modern C++ compilers no longer allow implicit conversion from integer to
enum values. Since the lv_obj_flag_t enum is intended to be used as a
bit flag, combining several values with the OR operator results in an
integer and not a enum value. Hence the compilation fails with an error:

  error: invalid conversion from ‘int’ to ‘lv_obj_flag_t’ [-fpermissive]